refactor(mobile): extract queue provider contexts and hooks
